Some orphans work
Removed `-Wno-orphans` from some modules.
Un-orphan some instances from `Ouroboros.Consensus.Util.Orphans`
