+++ ouroboros-consensus/changelog.d/orphans-javier.md
@@ -1,0 +1,3 @@
+### Non-Breaking
+
+- Un-orphan instances for `Condense` and `HeaderHash (Ticked l)`.
