+++ ouroboros-consensus/src/ouroboros-consensus/Ouroboros/Consensus/Protocol/PBFT/Crypto.hs
@@ -11,7 +11,6 @@
 {-# LANGUAGE TypeOperators          #-}
 {-# LANGUAGE UndecidableInstances   #-}
 
-{-# OPTIONS_GHC -Wno-orphans #-}
 module Ouroboros.Consensus.Protocol.PBFT.Crypto (
     PBftCrypto (..)
   , PBftMockCrypto
